requirements.txt
altgraph==0.17.5
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
macholib==1.16.4
modulegraph==0.19.7
packaging==26.0
py2app==0.28.9
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2026.0
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
requests==2.32.5
rumps==0.4.0
setuptools==80.10.2
urllib3==2.6.3
